RF_08.001.01 | Build history combining all Build history-related classes to one class (BuildhistoryTest)
