--- src/test/java/school/redrover/BuildHistory2Test.java
@@ -1,36 +1,0 @@
-package school.redrover;
-
-import org.openqa.selenium.By;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.Locale;
-
-import school.redrover.runner.BaseTest;
-
-public class BuildHistory2Test extends BaseTest {
-
-    @Test
-    public void testCheckDateAndMonthBuildHistory() {
-        Date systemDate = new Date();
-        SimpleDateFormat formatForDateNow = new SimpleDateFormat("MMM d", Locale.ENGLISH);
-        String dateNow = formatForDateNow.format(systemDate);
-
-        getDriver().findElement(By.className("content-block__link")).click();
-
-        getDriver().findElement(By.cssSelector(".jenkins-input")).sendKeys("Test");
-        getDriver().findElement(By.className("hudson_model_FreeStyleProject")).click();
-        getDriver().findElement(By.cssSelector("#ok-button")).click();
-        getDriver().findElement(By.xpath("//button[@formnovalidate='formNoValidate']")).click();
-
-        getDriver().findElement(By.cssSelector("a[href='/job/Test/build?delay=0sec']")).click();
-        if (dateNow.length() == 5) {
-            Assert.assertEquals(getWait10().until(ExpectedConditions.presenceOfElementLocated(By.xpath("//a[@class='model-link inside build-link']"))).getText().substring(0, 5), dateNow);
-        } else {
-            Assert.assertEquals(getWait10().until(ExpectedConditions.presenceOfElementLocated(By.xpath("//a[@class='model-link inside build-link']"))).getText().substring(0, 6), dateNow);
-        }
-    }
-}
